+++ Natom.Gestion.WebApp.Clientes.Backend.Entities/Model/OrdenDePedido.cs
@@ -67,7 +67,7 @@ namespace Natom.Gestion.WebApp.Clientes.Backend.Entities.Model
 		[ForeignKey("MarcoEntregaUsuarioId")]
 		public Usuario MarcoEntregaUsuario { get; set; }
 
-		public decimal PesoTotalEnGramos { get; set; }
+		public int PesoTotalEnGramos { get; set; }
 		public decimal? MontoTotal { get; set; }
 
 		public string MedioDePago { get; set; }
